Refactor model files
