+++ src/main/kotlin/astminer/common/model/StorageModel.kt
@@ -1,4 +1,4 @@
-package astminer.common
+package astminer.common.model
 
 
 /**
