+++ src/main/kotlin/astminer/examples/AllCppFiles.kt
@@ -2,7 +2,7 @@
 
 package astminer.examples
 
-import astminer.common.LabeledPathContexts
+import astminer.common.model.LabeledPathContexts
 import astminer.parse.cpp.FuzzyCppParser
 import astminer.paths.PathMiner
 import astminer.paths.PathRetrievalSettings
