+++ src/main/kotlin/astminer/parse/cpp/FuzzyMethodSplitter.kt
@@ -1,6 +1,7 @@
 package astminer.parse.cpp
 
 import astminer.common.*
+import astminer.common.model.*
 
 class FuzzyMethodSplitter : TreeMethodSplitter<FuzzyNode> {
