+++ src/test/kotlin/astminer/paths/PathWorkerTestUtil.kt
@@ -1,7 +1,7 @@
 package astminer.paths
 
-import astminer.common.ASTPath
-import astminer.common.Node
+import astminer.common.model.ASTPath
+import astminer.common.model.Node
 import astminer.common.postOrder
 import astminer.parse.antlr.SimpleNode
 import org.junit.Assert
